requirements.txt
Django==4.2.10
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
gunicorn==21.2.0